Count ways to divide a circle by chords
t -> no. of testcases
n -> no. of chords to be used
a -> no. of points on the circle used (2*n)

LOGIC -> Two ways are different if there exists a chord which is present in one way and not in other. When 2 points are joined to form a chord, two distinct sets on either side are formed. Let points on one side are j, and total is 2*n, so points on other side are 2*n-j-2 (2 is subtracted because ---> first chord uses 2 points). If we draw a chord from a point in first set to a point in second point, it will surely intersect the chord we’ve just drawn and hence will contradict the non-intersecting nature of chords as given in question.
An alternate way of solving this can be using the formula for finding nth Catalan number (These are a sequence of natural numbers that occur in many interesting counting problems like this one)