MinimumAddition
You are given an array A[]  of N positive numbers. You are also given Q  queries. Each query contains two integers l, r(l<=r). For each query, determine the sum of all values that must be added to each number in range  to  such that bitwise AND of all numbers in the range is greater than 0.

You are required to minimize the total value that must be added. The value you must add to numbers is not necessarily equal for all numbers.

Note: Each query is independent of each other, that is, for each query, you must treat the range over the initial input array.